requirements.txt
python-dotenv
pandas
numpy
requests
pymongo[srv]==3.12
pymongo
flask
certifi
scikit-learn

##-e .
